\name{datageneration.midas}
\alias{datageneration.midas}
\title{
datageneration.midas
}
\description{
Determine data frequency.
}
\usage{
datageneration.midas(DateVec)
}
\arguments{
  \item{DateVec}{vector of dates.}
}
\details{}
\value{}
\references{}
\author{J.S.}
\note{Description to be updated.}
\examples{}
