\name{diff.time.mf}
\alias{diff.time.mf}
\title{
TBA
}
\description{
TBA
}
\usage{
diff.time.mf(time1, time2, origin, units = c("auto", "secs", "mins", "hours", "days", "weeks"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{time1}{}
  \item{time2}{}
  \item{origin}{}
  \item{units}{}
}
\details{}
\value{}
\references{}
\author{J.S.}
\note{Description to be updated.}
\examples{}