\name{midas.adl}
\alias{midas.adl}
\title{
TBA
}
\description{
TBA
}
\usage{
midas.adl(data.y, data.ydate, data.x, data.xdate, est.start,est.end, horizon, x.lag, y.lag, polynomial, method, disp.flag,num.evals=10000,num.coef=5)
}
%- maybe also 'usage' for other objects documented here.
\details{}
\value{}
\references{}
\author{J.S.}
\note{Description to be updated.}
\examples{}