\name{midas.optimization}
\alias{midas.optimization}
\title{
midas.optimization
}
\description{
Determine data frequency.
}
\usage{
midas.optimization(DateVec)
}
\arguments{
  \item{DateVec}{vector of dates.}
}
\details{}
\value{}
\references{}
\author{J.S.}
\note{Description to be updated.}
\examples{}
