\name{MIDASLec-package}
\alias{MIDASLec-package}
\alias{MIDASLec}
\docType{package}
\title{
\packageTitle{MIDASLec}
}
\description{
\packageDescription{MIDASLec}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MIDASLec}
\packageIndices{MIDASLec}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MIDASLec}

Maintainer: \packageMaintainer{MIDASLec}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
