\name{mixed.freq.data}
\alias{mixed.freq.data}
\title{
mixed.freq.data
}
\description{
TBA
}
\usage{
mixed.freq.data(data.y, data.ydate, data.x, data.xdate, x.lag, y.lag, horizon, est.start, est.end, disp.flag = TRUE)
}

\arguments{
  \item{data.y}{}
  \item{data.ydate}{}
  \item{data.x}{}
  \item{data.xdate}{}
  \item{x.lag}{}
  \item{y.lag}{}
  \item{horizon}{}
  \item{est.start}{}
  \item{est.end}{}
  \item{disp.flag}{}
}
\details{}
\value{}
\references{}
\author{J.S.}
\note{Description to be updated.}
\examples{}