\name{mixed.freq.data.single}
\alias{mixed.freq.data.single}
\title{
mixed.freq.data.single
}
\description{
TBA
}
\usage{
mixed.freq.data.single(data.refdate, data.x, data.xdate, x.lag, horizon, est.start, est.end, disp.flag = TRUE)
}

\arguments{
  \item{data.refdate}{}
  \item{data.x}{}
  \item{data.xdate}{}
  \item{x.lag}{}
  \item{horizon}{}
  \item{est.start}{}
  \item{est.end}{}
  \item{disp.flag}{}
}
\details{}
\value{}
\references{}
\author{J.S.}
\note{Description to be updated.}
\examples{}