\name{MIDASLec-package}
\alias{MIDASLec-package}
\alias{MIDASLec}
\docType{package}
\title{
\packageTitle{MIDASLec}
}
\description{
\packageDescription{MIDASLec}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MIDASLec}
\packageIndices{MIDASLec}
}
\author{
\packageAuthor{MIDASLec}

Maintainer: \packageMaintainer{J.S.}
}
\references{
\keyword{ package }
\seealso{
}
\examples{
TBA
}
